Restricting access to some internal classes so that they will not be used by consumers of the sdk.
