+++ src/test/java/com/spectralogic/ds3client/ConnectionFixture.java
@@ -1,4 +1,4 @@
-package com.spectralogic.ds3client.fixtures;
+package com.spectralogic.ds3client;
 
 import com.spectralogic.ds3client.networking.ConnectionDetails;
 import com.spectralogic.ds3client.models.Credentials;
@@ -10,6 +10,6 @@ public class ConnectionFixture {
     }
 
     public static ConnectionDetails getConnection(final int port) {
-        return ConnectionDetails.builder("localhost:" + port, new Credentials("id", "key")).build();
+        return ConnectionDetailsImpl.builder("localhost:" + port, new Credentials("id", "key")).build();
     }
 }
